project/feature_extracion_vm_prof/requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
astunparse==1.6.3
black==21.4b2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.26
detectron2==0.6+cu113
filelock==3.4.2
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
huggingface-hub==0.2.1
hydra-core==1.1.1
idna==3.3
imantics==0.1.12
importlib-metadata==4.10.0
importlib-resources==5.4.0
install==1.3.5
iopath==0.1.9
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.22.0
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
Pillow==9.0.0
portalocker==2.3.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pydot==1.4.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.1
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
scipy==1.7.3
six==1.16.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
timm==0.4.12
tokenizers==0.10.3
toml==0.10.2
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.62.3
transformers==4.15.0
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
xmljson==0.2.1
yacs==0.1.8
zipp==3.7.0
